+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 - Allow enabling CORS using `--allow-cors <origin>` (#101)
 
+- Some minor UI improvements to the balance overview (soon to be replaced entirely)
+
 ## 0.2.8 - 2019-06-09
 
 - Security: Make npm-shrinkwrap actually ship with the npm package ([5827098ec](https://github.com/shesek/spark-wallet/commit/5827098ecc301e449e58e6077b0abf4e63aa9635))